+++ cocos/3d/CCObjLoader.cpp
@@ -477,7 +477,7 @@ namespace tinyobj {
                 char namebuf[TINYOBJ_SSCANF_BUFFER_SIZE];
                 token += 7;
 #ifdef _MSC_VER
-                sscanf_s(token, "%s", namebuf, _countof(namebuf));
+                sscanf_s(token, "%s", namebuf, unsigned(_countof(namebuf)));
 #else
                 sscanf(token, "%s", namebuf);
 #endif
@@ -785,7 +785,7 @@ namespace tinyobj {
                 char namebuf[TINYOBJ_SSCANF_BUFFER_SIZE];
                 token += 7;
 #ifdef _MSC_VER
-                sscanf_s(token, "%s", namebuf, _countof(namebuf));
+                sscanf_s(token, "%s", namebuf, unsigned(_countof(namebuf)));
 #else
                 sscanf(token, "%s", namebuf);
 #endif
@@ -814,7 +814,7 @@ namespace tinyobj {
                 char namebuf[TINYOBJ_SSCANF_BUFFER_SIZE];
                 token += 7;
 #ifdef _MSC_VER
-                sscanf_s(token, "%s", namebuf, _countof(namebuf));
+                sscanf_s(token, "%s", namebuf, unsigned(_countof(namebuf)));
 #else
                 sscanf(token, "%s", namebuf);
 #endif
@@ -880,7 +880,7 @@ namespace tinyobj {
                 char namebuf[TINYOBJ_SSCANF_BUFFER_SIZE];
                 token += 2;
 #ifdef _MSC_VER
-                sscanf_s(token, "%s", namebuf, _countof(namebuf));
+                sscanf_s(token, "%s", namebuf, unsigned(_countof(namebuf)));
 #else
                 sscanf(token, "%s", namebuf);
 #endif
